+++ packages/amplify-graphql-relational-transformer/src/graphql-has-many-transformer.ts
@@ -14,6 +14,7 @@ import {
   getFieldsNodes,
   getRelatedType,
   getRelatedTypeIndex,
+  validateDisallowedDataStoreRelationships,
   validateModelDirective,
   validateRelatedModelDirective,
 } from './utils';
@@ -84,4 +85,5 @@ function validate(config: HasManyDirectiveConfiguration, ctx: TransformerContext
   config.relatedType = getRelatedType(config, ctx);
   config.connectionFields = [];
   validateRelatedModelDirective(config);
+  validateDisallowedDataStoreRelationships(config, ctx);
 }