+++ packages/amplify-util-mock/src/api/lambda-arn-to-config.ts
@@ -8,11 +8,10 @@ import { ProcessedLambdaFunction } from '../CFNParser/stack/types';
 /**
  * Attempts to match an arn object against the array of lambdas configured in the project
  */
-export const lambdaArnToConfig = (context: $TSContext, arn: $TSAny): Promise<ProcessedLambdaFunction> => {
-  const version = getTransformerVersion(context);
+export const lambdaArnToConfig = async (context: $TSContext, arn: $TSAny): Promise<ProcessedLambdaFunction> => {
+  const version = await getTransformerVersion(context);
   const doclink = getGraphQLTransformerFunctionDocLink(version);
-  const errorSuffix =
-    `\nSee ${doclink} for information on how to configure Lambda resolvers.`;
+  const errorSuffix = `\nSee ${doclink} for information on how to configure Lambda resolvers.`;
   let searchString = '';
   if (typeof arn === 'string') {
     searchString = arn;